fix: randoms up to and over 4000000000, implement aks
- https://en.wikipedia.org/wiki/AKS_primality_test
- https://stackoverflow.com/a/1801446/5306470

+++ Primes Fun/Controller/Random/RandomResultsViewController.swift
@@ -20,7 +20,7 @@ class RandomResultsViewController: UIViewController {
     
     // MARK: Properties
     
-    var number: Int!
+    var number: Int64!
     
     let positiveSound: Int = 1023
     let negativeSound: Int = 1257
